+++ torch/csrc/jit/passes/onnx/peephole.h
@@ -1,0 +1,7 @@
+#include "torch/csrc/jit/ir.h"
+
+namespace torch { namespace jit {
+
+void PeepholeOptimizeONNX(std::shared_ptr<Graph>& graph);
+
+}}
