+++ torch/onnx/__init__.py
@@ -96,9 +96,12 @@ def _export(model, args, f, export_params=True, verbose=False, training=False):
         raise RuntimeError("state_dict changed after running the tracer; "
                            "something weird is happening in your model!")
 
+    # TODO: write a helper function
     torch._C._jit_pass_peephole(trace)
     torch._C._jit_pass_lint(trace)
     torch._C._jit_pass_onnx(trace)
+    torch._C._jit_pass_lint(trace)
+    torch._C._jit_pass_onnx_peephole(trace)
     torch._C._jit_pass_lint(trace)
     torch._C._jit_pass_dce(trace)
     torch._C._jit_pass_lint(trace)
